Extract URL state management to hook
